refactor: TLS 强制v1.3加密、代理管理器完善热更新处理、TCP接口更新完善

+++ console/src/main/java/com/xiaoniucode/etp/server/web/service/impl/ProxyServiceImpl.java
@@ -105,7 +105,6 @@ public class ProxyServiceImpl implements ProxyService {
 
         //生成代理配置信息
         String proxyId = uidGenerator.getUIDAsString();
-
         ProxyDO proxyDO = proxyConvert.toDO(param, proxyId);
         proxyDO.setSourceType(ProxySourceType.MANUAL);
         BandwidthSaveParam bandwidth = param.getBandwidth();
@@ -204,6 +203,7 @@ public class ProxyServiceImpl implements ProxyService {
                 proxyDO.getAgentId(), param.getName(), proxyId)) {
             throw new BizException("该客户端下已存在同名代理名称: " + param.getName());
         }
+        DomainType existsDomainType = proxyDO.getDomainType();
         //1.基本信息
         proxyConvert.updateDO(param, proxyDO);
         proxyDO.setSourceType(ProxySourceType.MANUAL);
@@ -223,7 +223,6 @@ public class ProxyServiceImpl implements ProxyService {
         //4.负载均衡
         if (proxyDO.getDeploymentMode().isCluster()) {
             LoadBalanceParam loadBalanceParam = param.getLoadBalance();
-
             LoadBalanceDO loadBalanceDO = loadBalanceRepository.findById(proxyId)
                     .orElseGet(() -> {
                         LoadBalanceDO newLB = new LoadBalanceDO();
@@ -237,16 +236,56 @@ public class ProxyServiceImpl implements ProxyService {
             loadBalanceRepository.deleteById(proxyId);
         }
         //6.传输
-        transportRepository.deleteById(proxyId);
-        TransportDO transportDO = transportConvert.toDO(param.getTransport(), proxyId);
-        transportRepository.save(transportDO);
-        //todo 7.HTTP域名信息
-        proxyDomainRepository.deleteByProxyId(proxyId);
-
-//        if (!CollectionUtils.isEmpty(boundDomains)) {
-//            List<HttpProxyDomainDO> domainList = proxyDomainConvert.toDOList(boundDomains, proxyId);
-//            proxyDomainRepository.saveAll(domainList);
-//        }
+        transportRepository.save(transportConvert.toDO(param.getTransport(), proxyId));
+        //7.HTTP域名信息
+        if (!(existsDomainType.getCode().equals(param.getDomainType()) && existsDomainType.isAuto())) {
+            proxyDomainRepository.deleteByProxyId(proxyId);
+        }
+        String baseDomain = appConfig.getBaseDomain();
+        DomainType domainType = proxyDO.getDomainType();
+        if (domainType.isAuto() && !existsDomainType.isAuto()) {
+            String domain = domainGenerator.generateSubdomain(baseDomain);
+            proxyDomainRepository.save(new ProxyDomainDO(proxyId, domain, baseDomain, domainType));
+        } else if (domainType.isCustomDomain()) {
+            Set<String> domains = param.getDomains();
+            List<ProxyDomainDO> existsList = proxyDomainRepository.findByFullDomainIn(domains);
+            if (!existsList.isEmpty()) {
+                String existDomains = existsList.stream()
+                        .map(ProxyDomainDO::getDomain)
+                        .collect(Collectors.joining(", "));
+                throw new BizException("以下域名已被使用: " + existDomains);
+            }
+            List<ProxyDomainDO> list = domains.stream()
+                    .map(domain -> new ProxyDomainDO(
+                            proxyId,
+                            domain,
+                            null,
+                            domainType
+                    ))
+                    .toList();
+            proxyDomainRepository.saveAll(list);
+        } else if (domainType.isSubdomain()) {
+            Set<String> prefixes = param.getDomains();
+            List<String> fullDomains = prefixes.stream()
+                    .map(prefix -> prefix + "." + baseDomain)
+                    .toList();
+            List<ProxyDomainDO> existsList = proxyDomainRepository.findByFullDomainIn(fullDomains);
+            if (!existsList.isEmpty()) {
+                String existDomains = existsList.stream()
+                        .map(ProxyDomainDO::getDomain)
+                        .collect(Collectors.joining(", "));
+                throw new BizException("以下子域名已被使用: " + existDomains);
+            }
+            List<ProxyDomainDO> list = prefixes.stream()
+                    .map(prefix -> new ProxyDomainDO(
+                            proxyId,
+                            prefix,
+                            baseDomain,
+                            domainType
+                    ))
+                    .toList();
+            proxyDomainRepository.saveAll(list);
+        }
         transactionHelper.afterCommit(() ->
                 proxyManager.reconcile(proxyAssembler.toProxyConfig(proxyDO)));
         logger.debug("HTTP代理更新成功：{}", proxyDO.getName());
@@ -427,6 +466,7 @@ public class ProxyServiceImpl implements ProxyService {
         transportRepository.deleteById(proxyId);
         TransportDO transportDO = transportConvert.toDO(param.getTransport(), proxyId);
         transportRepository.save(transportDO);
+        transactionHelper.afterCommit(()-> proxyManager.reconcile(proxyAssembler.toProxyConfig(proxyDO)));
         logger.debug("TCP 代理更新成功：{}", proxyDO.getName());
     }
